Implement deep equality hash checks
Ben pointed out that the current hash implementation is simply
incrementing a number. In trivial tests (that I naively ran while
testing the original change), things worked fine because the sets and
maps were under a threshold that would cause the hashing to even occur.

Exceeding ~30 items caused hashing to take place, and then the ++ID
approach naturally didn't work. This implements a deep equality hash
check for vanilla JS objects and adds a couple of tests to help
detect this issue.